refactor(mgmt): combine espoo audience management commands
For simplicity, combine Espoo audience management commands. They
probably change for the same reasons and at the same time so it's best
to combine them into one command.
